Ref(axios): Move config to a separate file
This allows for using the axios instance in places where the Vue
instance (this._vm) is not available, like the getter functions in Vuex.
The other option is to import the Vue instance separately inside the
vuex module (assuming axios instance has already been placed in the
Vue.prototype before registering Vuex with Vue), which is not ideal.