Fix: Force user to clear search term using the clear button
This is a quick fix to avoid an invalid UI state where the user manually
clears the search field - which clears the search results but the result
cache remains in vuex. Therefore, any subsequent keypresses will
immediately show the results of the previous search.
Forcing the user to use the clear button will clear the search results
as well as the cached data in vuex.
Some might complain this is a hacky solution, but I think this is enough
for now. If this ends up confusing a lot of people then an alternative
solution may be implemented.